+++ server-db/src/main/resources/db/changelog/changes/010-key_material_share_add_column_expiry_time_adjusted.sql
@@ -1,0 +1,4 @@
+-- liquibase formatted sql
+-- changeset cdoc2_capsule_add_column_expiry_time_adjusted:6
+ALTER TABLE key_material_share
+ADD expiry_time_adjusted BOOLEAN DEFAULT FALSE;