+++ shares-server/src/main/java/ee/cyber/cdoc2/server/Constants.java
@@ -1,0 +1,9 @@
+package ee.cyber.cdoc2.server;
+
+public final class Constants {
+    private Constants() {
+    }
+
+    public static final String X_EXPIRY_TIME_HEADER = "x-expiry-time";
+    public static final String X_EXPIRY_TIME_ADJUSTED = "x-expiry-time-adjusted";
+}